api_backend/requirements.txt
django
djangorestframework
openai
gunicorn
